requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.25.6
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
base58==2.1.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
communex==0.1.16
cytoolz==0.12.3
decorator==5.1.1
distro==1.9.0
ecdsa==0.19.0
eth-hash==0.7.0
eth-keys==0.5.1
eth-typing==4.2.1
eth-utils==2.3.1
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.110.0
filelock==3.13.4
frozenlist==1.4.1
fsspec==2024.3.1
gensim==4.3.2
gnureadline==8.1.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
ipython==8.23.0
jedi==0.19.1
keylimiter==0.1.5
loguru==0.7.2
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.2.0
multidict==6.0.5
numpy==1.26.4
openai==1.23.5
packaging==24.0
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
py-bip39-bindings==0.1.11
py-ed25519-zebra-bindings==1.0.1
py-sr25519-bindings==0.2.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.17.2
PyNaCl==1.5.0
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
rich==13.7.1
safetensors==0.4.3
scalecodec==1.2.8
scipy==1.13.0
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
stack-data==0.6.3
starlette==0.36.3
substrate-interface==1.7.8
tokenizers==0.19.1
toolz==0.12.1
tqdm==4.66.2
traitlets==5.14.3
transformers==4.40.1
typeguard==4.2.1
typer==0.9.4
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
wcwidth==0.2.13
websocket-client==1.8.0
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
